% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-frequency.R
\name{bigrams}
\alias{bigrams}
\title{Bigrams}
\usage{
bigrams(text)
}
\arguments{
\item{text}{Sequence of items, as an iterator.}
}
\description{
Generate bigrams.
}
\examples{
bigrams(list("Generate", "some", "bigrams"))

}
