% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{books}
\alias{books}
\title{Book}
\usage{
books(text = "text1")
}
\arguments{
\item{text}{The text to return.}
}
\description{
Access data useful to follow along the \href{https://www.nltk.org/book}{NLTK Book}.
}
\examples{
mody_dick <- books()
mody_dick$concordance("monstrous")
mody_dick$similar("monstrous")

}
