% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{brown_corpus}
\alias{brown_corpus}
\title{Brown Corpus}
\usage{
brown_corpus()
}
\value{
Return NLTK's Brown corpus as Python object.
}
\description{
The Brown Corpus was the first million-word electronic corpus of English, 
created in 1961 at Brown University. This corpus contains text from 500 
sources, and the sources have been categorized by genre, such as news, 
editorial, and so on. 1.1 gives an example of each genre (for a 
complete list, see \url{http://icame.uib.no/brown/bcm-los.html}).
}
\examples{
\dontrun{
brown <- brown_corpus()
brown$categories() 
}

}
