% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_accuracy}
\alias{classify_accuracy}
\title{Classification Accuracy}
\usage{
classify_accuracy(classifier, test)
}
\arguments{
\item{classifier}{A classifier (model).}

\item{test}{The test set.}
}
\description{
Assess classification Accuracy.
}
