% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{corpus_reader}
\alias{corpus_reader}
\title{Corpus Reader}
\usage{
corpus_reader(root, pattern = ".*")
}
\arguments{
\item{root}{The root directory.}

\item{pattern}{Pattern to match file neames against.}
}
\description{
Read collection of text files as corpus.
}
