% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{download_datasets}
\alias{download_datasets}
\title{Resources}
\usage{
download_datasets(resource = "all")
}
\arguments{
\item{resource}{Resource to download, defaults to \code{all}, set to \code{NULL} to get prompted.}
}
\description{
Download required resources.
}
\examples{
\dontrun{download_datasets("punkt")}

}
