% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{first_names}
\alias{first_names}
\title{First Names}
\usage{
first_names(to_r = FALSE)
}
\arguments{
\item{to_r}{Whether to return results in R tidy format.}
}
\description{
First names to build a gender classifier.
}
\examples{
\dontrun{
first_names(to_r = TRUE)
}

}
