% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gutenberg}
\alias{gutenberg}
\alias{gutenberg_files}
\alias{gutenberg_raw}
\alias{gutenberg_words}
\alias{gutenberg_sents}
\title{Gutenberg}
\usage{
gutenberg_files(to_r = FALSE)

gutenberg_raw(file, to_r = FALSE)

gutenberg_words(file, to_r = FALSE)

gutenberg_sents(file, to_r = FALSE)
}
\arguments{
\item{to_r}{Whether to return results in R tidy format.}

\item{file}{A Gutenberg file as returned by \code{gutenberg_files}.}
}
\description{
Small selection of texts from the Project Gutenberg electronic text archive.
}
\examples{
\dontrun{
files <- gutenberg_files(to_r = TRUE)
words <- gutenberg_words(sample(files, 1)) # get random text
text <- nltk_text(words) # convert
text$concordance("surprize")
}

}
