% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{inaugural}
\alias{inaugural}
\alias{inaugural_files}
\alias{inaugural_words}
\alias{inaugural_sents}
\alias{inaugural_raw}
\title{Inaugural Address Corpus}
\usage{
inaugural_files(to_r = FALSE)

inaugural_words(file, to_r = FALSE)

inaugural_sents(file, to_r = FALSE)

inaugural_raw(file, to_r = FALSE)
}
\arguments{
\item{to_r}{Whether to return results in R tidy format.}

\item{file}{An Inaugural address as returned by \code{inaugural_files}.}
}
\description{
Corpus of Inaugural Addresses of US Presidents.
}
\examples{
\dontrun{
files <- inaugural_files(to_r = TRUE)
inaugural_sents(sample(files, 1)) 
}

}
