% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{install_nltk}
\alias{install_nltk}
\title{Install Dependencies}
\usage{
install_nltk(method = "auto", conda = "auto")
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically 
finds a method that will work in the local environment. 
Change the default to force a specific installation method. 
Note that the "virtualenv" method is not available on Windows.}

\item{conda}{Path to conda executable (or "auto" to find conda using the
\code{PATH} and other conventional install locations).}
}
\description{
Install the Natural Language dependency.
}
\examples{
\dontrun{install_nltk()}

}
