% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{movie_reviews}
\alias{movie_reviews}
\alias{movie_reviews_files}
\alias{movie_reviews_words}
\alias{movie_reviews_categories}
\alias{movie_reviews_sents}
\alias{movie_reviews_raw}
\title{Movie Reviews Corpus}
\usage{
movie_reviews_files(..., to_r = FALSE)

movie_reviews_words(file, to_r = FALSE)

movie_reviews_categories(file = NULL, to_r = FALSE)

movie_reviews_sents(file, to_r = FALSE)

movie_reviews_raw(file, to_r = FALSE)
}
\arguments{
\item{...}{Any other options such as category.}

\item{to_r}{Whether to return results in R tidy format.}

\item{file}{A file containing movie reviews as returned by \code{movie_reviews_files}.}
}
\description{
Corpus of movie reviews.
}
\examples{
\dontrun{
files <- movie_reviews_files(to_r = TRUE)
movie_reviews_words(sample(files, 1)) 

# by category
(movie_reviews_categories())
positive_reviews <- movie_reviews_files("pos", to_r = TRUE)
}

}
