% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokeniser.R
\name{ne_chunk}
\alias{ne_chunk}
\title{Named Entity Extraction}
\usage{
ne_chunk(pos)
}
\arguments{
\item{pos}{POS tagged to tokenize.}
}
\description{
Named Entity Extraction
}
\examples{
\dontrun{
tokens <- word_tokenize("This is an R package.") 
tagged <- pos_tag(tokens)
ne_chunk(tagged)
}

}
