% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokeniser.R
\name{nltk_ne_chunk}
\alias{nltk_ne_chunk}
\title{Named Entity Extraction}
\usage{
nltk_ne_chunk(pos)
}
\arguments{
\item{pos}{POS tagged to tokenize.}
}
\description{
Named Entity Extraction
}
\examples{
\dontrun{
tokens <- nltk_word_tokenize("This is an R package.") 
tagged <- nltk_pos_tag(tokens)
nltk_ne_chunk(tagged)
}

}
