% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{nltk_text}
\alias{nltk_text}
\title{As text}
\usage{
nltk_text(text)
}
\arguments{
\item{text}{Text to convert.}
}
\description{
As NLTK text object.
}
\examples{
\dontrun{
files <- gutenberg_files(to_r = TRUE)
words <- gutenberg_words(sample(files, 1)) # get random text
text <- nltk_text(words) # convert
text$concordance("surprize")
}

}
