% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokeniser.R
\name{nltk_word_tokenize}
\alias{nltk_word_tokenize}
\title{Tokenizers Words}
\usage{
nltk_word_tokenize(sentence)
}
\arguments{
\item{sentence}{Sentence to tokenize.}
}
\description{
Word tokenizer
}
\examples{
\dontrun{
nltk_word_tokenize("This is an R package.") 
}

}
