% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokeniser.R
\name{pos_tag}
\alias{pos_tag}
\title{POS Tagger}
\usage{
pos_tag(tokens, to_r = TRUE)
}
\arguments{
\item{tokens}{Sentence to tokenize.}

\item{to_r}{Wehther to return results in tidy format.}
}
\description{
Parts of speech tagger
}
\examples{
\dontrun{
tokens <- word_tokenize("This is an R package.") 
pos_tag(tokens)
}

}
