% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{reuters}
\alias{reuters}
\alias{reuters_files}
\alias{reuters_categories}
\alias{reuters_raw}
\alias{reuters_words}
\title{Reuters}
\usage{
reuters_files(category = NULL, to_r = FALSE)

reuters_categories(..., to_r = FALSE)

reuters_raw(..., to_r = FALSE)

reuters_words(..., to_r = FALSE)
}
\arguments{
\item{category}{A category or a list of categories, see examples.}

\item{to_r}{Whether to return results in R tidy format.}

\item{...}{Additional parameters such as \code{categories}, see examples.}
}
\description{
The Reuters Corpus contains 10,788 news documents totaling 1.3 million words. 
The documents have been classified into 90 topics, and grouped into two sets, 
called "training" and "test"; thus, the text with fileid 'test/14826' is a 
document drawn from the test set.
}
\examples{
\dontrun{
files <- reuters_files(to_r = TRUE)
categories <- reuters_categories() # all categories
reuters_categories(sample(files, 1)) 
reuters_words(categories = "barley") 
reuters_files("barley") 
}

}
