% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{shuffle}
\alias{shuffle}
\title{Shuffle}
\usage{
shuffle(x)
}
\arguments{
\item{x}{A dataset to shuffle.}
}
\value{
A Python object.
}
\description{
Shuffle a dataset.
}
\examples{
\dontrun{
letters \%>\% 
  apply(1, as.list) \%>\% 
  shuffle()
}

}
