% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{train_bayes_classifier}
\alias{train_bayes_classifier}
\title{Naive Bayes Classifer}
\usage{
train_bayes_classifier(train)
}
\arguments{
\item{train}{The training set.}
}
\description{
Train a naive Bayes classifier.
}
