% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokeniser.R
\name{word_tokenize}
\alias{word_tokenize}
\title{Tokenizers Words}
\usage{
word_tokenize(sentence)
}
\arguments{
\item{sentence}{Sentence to tokenize.}
}
\description{
Word tokenizer
}
\examples{
\dontrun{
word_tokenize("This is an R package.") 
}

}
