% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{download_datasets}
\alias{download_datasets}
\title{Resources}
\usage{
download_datasets(resource = NULL)
}
\arguments{
\item{resource}{Resource to download.}
}
\description{
Download required resources.
}
\section{Resources}{

\itemize{
  \item{\code{punkt}}
  \item{\code{averaged_perceptron_tagger}}
}
}

\examples{
\dontrun{download_datasets("punkt")}

}
