% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{nltk_download}
\alias{nltk_download}
\title{Resources}
\usage{
nltk_download(resource)
}
\arguments{
\item{resource}{Resource to download.}
}
\description{
Download required resources.
}
\examples{
\dontrun{nltk_download("punkt")}

}
