% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokeniser.R
\name{nltk_pos_tag}
\alias{nltk_pos_tag}
\title{POS Tagger}
\usage{
nltk_pos_tag(tokens, tidy = TRUE)
}
\arguments{
\item{tokens}{Sentence to tokenize.}

\item{tidy}{Wehther to return results in tidy format.}
}
\description{
Parts of speech tagger
}
\examples{
\\donrun{
tokens <- nltk_word_tokenize("This is an R package.") 
nltk_pos_tag(tokens)
}

}
