% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokeniser.R
\name{nltk_pos_tag}
\alias{nltk_pos_tag}
\title{Tagger}
\usage{
nltk_pos_tag(tokens)
}
\arguments{
\item{tokens}{Sentence to tokenize.}
}
\description{
Tagger
}
\examples{
\\donrun{
tokens <- nltk_word_tokenize("This is an R package.") 
nltk_pos_tag(tokens)
}

}
